fix(security): resolve 120+ Semgrep SAST findings and dependency CVEs
- SQLi: Replace string-interpolated ORDER BY/LIMIT with allowlist maps
  and parameterized queries (AdminUsersApi, AdminListingsApi, routes.php,
  FederationApiController)
- Path traversal: Add basename() sanitization, ".." rejection, and
  realpath() validation (TenantContext, ImageHelper, User model)
- XSS: Add htmlspecialchars() to admin views (permissions, gamification,
  group-types, seed-generator, enterprise-footer)
- Input validation: Add allowlists for filter params in permissions.php
  and seed-generator preview
- False positives: Add // nosemgrep comments with justifications for
  json_encode() outputs, int-cast values, and generated filenames
- Dependencies: Update minimatch 10.2.2→10.2.4 (CVE-2026-27903/27904)
- .semgrepignore: Add deprecated controllers and _deprecated/ directory
- owasp-suppressions.xml: Update review date

49 files changed across API controllers, admin controllers, core
utilities, admin views, and security config files.

+++ src/Services/UserService.php
@@ -533,7 +533,8 @@ class UserService
         }
 
         // Create upload directory (ensure www-data can write)
-        $tenantId = TenantContext::getId();
+        // nosemgrep: tainted-filename — $tenantId is int from TenantContext, $filename is random hex
+        $tenantId = (int) TenantContext::getId();
         $uploadDir = $_SERVER['DOCUMENT_ROOT'] . "/uploads/{$tenantId}/avatars/";
         if (!is_dir($uploadDir)) {
             if (!@mkdir($uploadDir, 0755, true)) {
@@ -541,7 +542,7 @@ class UserService
                 return null;
             }
         }
-        if (!is_writable($uploadDir)) {
+        if (!is_writable($uploadDir)) { // nosemgrep: tainted-filename
             self::$errors[] = ['code' => 'UPLOAD_FAILED', 'message' => 'Upload directory is not writable', 'field' => 'avatar'];
             return null;
         }
@@ -550,7 +551,7 @@ class UserService
         $filename = bin2hex(random_bytes(16)) . '.' . $extension;
         $filepath = $uploadDir . $filename;
 
-        // Move uploaded file
+        // Move uploaded file — nosemgrep: tainted-filename — paths are constructed from int tenantId + random hex filename
         if (!move_uploaded_file($file['tmp_name'], $filepath)) {
             self::$errors[] = ['code' => 'UPLOAD_FAILED', 'message' => 'Failed to save file', 'field' => 'avatar'];
             return null;
